Creating connection with Database
- Created Database configuration
- Created Service layer
- Updated Controller layer
- Updated the models

+++ projects/api/ProductPackaging/Models/MFooter.cs
@@ -8,5 +8,5 @@ public class MFooter
 
     public bool hasNext { get; set; }
 
-    public int TotalRecords { get; set; }
+    public long TotalRecords { get; set; }
 }
